Creating ID API Search
Originally, aimed to gather the 1st pages id range in order to create a modified binary search algorithm. However, the IDs are ordered weirdly for people (i.e. jumps from 91xxxxx to 93xxxxx) so that appears to be a dead end. A basic binary search algorithm will likely be much more flexible and accurate.